+++ js/venue-check.js
@@ -1,8 +1,10 @@
+/* venue-check.js v2.2.0-rc5 */
 /* eslint-disable no-console */
 /* eslint-disable camelcase */
 /* global venuecheck */
-( function( $ ) {
-	window.addEventListener( 'load', function() {
+jQuery( function( $ ) {
+	window.addEventListener( 'load', ( e ) => {
+		if ( venuecheck.debug ) console.log( 'load event fire', e );
 		//=====FORM MODIFIED=====//
 
 		//only trigger modified message on updating events
@@ -78,6 +80,7 @@
 			event_recurrences.push( event_recurrence );
 			const post_data = $( 'form#post' ).serialize();
 			venuecheck_disable_form();
+			if ( venuecheck.debug ) console.log( 'nonce: ' + venuecheck.nonce );
 			if ( typeof formVars[ 'is_recurring[]' ] !== 'undefined' ) {
 				if ( venuecheck.debug ) console.log( 'RECURRING EVENT' );
 				return $.ajax( {
@@ -86,7 +89,7 @@
 					dataType: 'json',
 					data: {
 						action: 'venuecheck_get_event_recurrences',
-						security: $( '#venuecheck_nonce' ).val(),
+						nonce: venuecheck.nonce,
 						post_data,
 					},
 				} )
@@ -170,12 +173,14 @@
 					const total_count = event_recurrences.length;
 					const batch_count = batchArray.indexOf( batch );
 					return p.then( function() {
+						if ( venuecheck.debug ) console.log( 'nonce: ' + venuecheck.nonce );
 						return $.ajax( {
 							type: 'POST',
 							dataType: 'json',
 							url: venuecheck.ajax_url,
 							data: {
 								action: 'venuecheck_check_venues',
+								nonce: venuecheck.nonce,
 								event_recurrences: batch,
 								postID,
 								total_count,
@@ -191,9 +196,9 @@
 								if ( batchArray.length > 1 ) {
 									venuecheck_check_venues_progress( percent_current, percent_end );
 								}
-								$( '#saved_tribe_venue' ).select2( {
-									disabled: true,
-								} );
+								$( '#saved_tribe_venue' )
+									.prop( 'readonly', true )
+									.addClass( 'readonly' );
 							},
 						} ).then( function( data ) {
 							venuecheck_conflicts = venuecheck_conflicts.concat( data );
@@ -378,9 +383,9 @@
 
 			$( '#venuecheck-conflicts-link' ).removeClass( 'venuecheck-disabled' );
 			$( '#publish' ).prop( 'disabled', false );
-			$( '#saved_tribe_venue' ).select2( {
-				disabled: false,
-			} );
+			$( '#saved_tribe_venue' )
+				.prop( 'readonly', false )
+				.removeClass( 'readonly' );
 			$( '#venuecheck-change-venue' ).hide();
 			$( 'body' ).addClass( 'venuecheck-venues' );
 			venuecheck_enable_form();
@@ -436,9 +441,9 @@
 			$( '#venuecheck-messages-container, #venuecheck-modified-publish, #venuecheck-modified' ).show();
 			$( '#venuecheck-report-container, #venuecheck-conflicts-report-count, #venuecheck-progress' ).hide();
 			$( '#publish' ).prop( 'disabled', true );
-			$( '#saved_tribe_venue' ).select2( {
-				disabled: true,
-			} );
+			$( '#saved_tribe_venue' )
+				.prop( 'readonly', true )
+				.addClass( 'readonly' );
 			$( '#venuecheck-change-venue' ).show();
 		}
 
@@ -453,9 +458,9 @@
 				$( '#venuecheck-report-container' ).hide();
 				$( '#venuecheck-conflicts-report-link' ).text( 'Show Details' );
 				//here
-				$( '#saved_tribe_venue' ).select2( {
-					disabled: false,
-				} );
+				$( '#saved_tribe_venue' )
+					.prop( 'readonly', false )
+					.removeClass( 'readonly' );
 			}
 		}
 
@@ -484,11 +489,9 @@
 		function venuecheck_disable_form() {
 			$( '#saved_tribe_venue' )
 				.prop( 'readonly', true )
+				.addClass( 'readonly' )
 				.addClass( 'venuecheck-preserve-disabled' );
 			$( '.tribe-datetime-block :input:disabled' ).addClass( 'venuecheck-preserve-disabled' );
-			$( '#saved_tribe_venue' ).select2( {
-				disabled: true,
-			} );
 			$( '#publish' ).prop( 'disabled', true );
 			$( '.tribe-datetime-block :input' ).prop( 'disabled', true );
 			$( '#tribe_events_event_details a' )
@@ -499,7 +502,9 @@
 		}
 
 		function venuecheck_enable_form() {
-			$( '#saved_tribe_venue:not(.venuecheck-preserve-disabled)' ).select2( 'readonly', false );
+			$( '#saved_tribe_venue:not(.venuecheck-preserve-disabled)' )
+				.prop( 'readonly', false )
+				.removeClass( 'readonly' );
 			$( '.tribe-datetime-block :input:not(.venuecheck-preserve-disabled)' ).prop( 'disabled', false );
 			$( '#saved_tribe_venue, .tribe-datetime-block :input' ).removeClass( 'venuecheck-preserve-disabled' );
 			$( '#publish' ).prop( 'disabled', false );
@@ -530,37 +535,49 @@
 			}
 		}
 
+		// Setup additional required HTML elements and classes
+		// It would be great to do this in PHP instead, but there is no way to modify the form
+
+		const $eventDetails = $( '#tribe_events_event_details' );
+		const $venuecheckVenueSection = $( '#event_tribe_venue tbody' ).first();
+		const $venueDropdown = $venuecheckVenueSection.find( 'tr.saved-linked-post > td' ).last();
 		//add venuecheck
-		$( '#tribe_events_event_details' ).addClass( 'venuecheck-update' );
+		$eventDetails.addClass( 'venuecheck-update' );
 
 		//add section
-		$( '#event_tribe_venue tbody:first' ).addClass( 'venuecheck-section' );
+		$venuecheckVenueSection.addClass( 'venuecheck-section' );
 
 		//add section label
-		$( '#event_tribe_venue > tbody > tr:first' ).before(
-			'<tr>' +
-				'<td colspan="2" class="venuecheck-section-label">' +
-				'Venue Check<i class="fas fa-spinner fa-pulse venuecheck-preload"></i>' +
-				'</td>' +
-				'</tr>'
-		);
+		$venuecheckVenueSection
+			.children( 'tr' )
+			.first()
+			.before(
+				'<tr>' +
+					'<td colspan="2" class="venuecheck-section-label">' +
+					'Venue Check<i class="fas fa-spinner fa-pulse venuecheck-preload"></i>' +
+					'</td>' +
+					'</tr>'
+			);
 
 		//add conflicts button
-		$( '#event_tribe_venue .venuecheck-section tr:nth-child(2) td:nth-child(2)' ).wrapInner(
-			'<div id="venuecheck-venue-select"></div>'
-		);
-		$( '#event_tribe_venue .venuecheck-section tr:nth-child(2) td:nth-child(2)' ).append(
-			'<a id="venuecheck-conflicts-button" class="button">Find available venues</a>'
-		);
+		$venueDropdown.wrapInner( '<div id="venuecheck-venue-select"></div>' );
+
+		$venueDropdown.append( '<a id="venuecheck-conflicts-button" class="button">Find available venues</a>' );
 
 		//add messages
-		$( '#event_tribe_venue .venuecheck-section tr:nth-child(2) td:nth-child(2)' ).append(
+		$venueDropdown.append(
 			'<div id="venuecheck-messages-container" style="display: none;"><div id="venuecheck-messages"></div></div>'
 		);
-		$( '#event_tribe_venue .venuecheck-section tr:nth-child(3)' ).after(
-			'<tr id="venuecheck-report-container" style="display: none;"><td colspan="2" id="venuecheck-report"></td>/tr>'
-		);
-		$( '#event_tribe_venue .venuecheck-section tr:nth-child(4)' ).after( '<tr class="venuecheck-spacer"><td></td><td></td></tr>' );
+
+		$venuecheckVenueSection
+			.find( 'tr.linked-post.venue' )
+			.first()
+			.after(
+				'<tr id="venuecheck-report-container" style="display: none;">' +
+					'<td colspan="2" id="venuecheck-report"></td>' +
+					'</tr>' +
+					'<tr class="venuecheck-spacer"><td></td><td></td></tr>'
+			);
 
 		//add conflicts link
 		$( '#event_tribe_venue .edit-linked-post-link' ).after(
@@ -642,9 +659,9 @@
 		*/
 
 		//disable venues dropdown
-		$( '#saved_tribe_venue' ).select2( {
-			disabled: true,
-		} );
+		$( '#saved_tribe_venue' )
+			.prop( 'readonly', true )
+			.addClass( 'readonly' );
 
 		// empty out the link to edit the venue item
 		$( '.edit-linked-post-link' ).html( '' );
@@ -655,4 +672,4 @@
 		venuecheck_show_hide_offsets();
 		venuecheck_show_hide_divider();
 	} ); //load
-} )( jQuery );
+} );